requirements.txt
pymongo==4.0.1
python-dotenv==0.19.2
python-telegram-bot==13.11
requests==2.27.1
tqdm==4.62.3
